backend/requirements.txt
Flask
psycopg2-binary # conexiunea la baza de date
